Updated the main function - see description.
The main function was reorganised to make use of the Sort class status boolean which indicated if the object array was sorted or not. A reset button was added enabling the user to start the sort process again and set a new sort type. The colours of the interface were also improved and button positions altered. Overall, the main function had some initial refactoring to help tidy the code.